Rename fubar to swiplwasmHandler
